requirements.txt
dnspython==1.14.0
funcsigs==1.0.2
mock==2.0.0
pbr==1.10.0
PyYAML==3.11
requests==2.10.0
six==1.10.0
slacker==0.9.16
